Ran a better training session with a population of 100 for about 7 hrs. We made it through 73 generations. Tested the best individual of the last generation and got incredible results against out most basic heuristic. The beginning and mid game were awesome. The solver was making very good moves, multiple captures consecutively, smart sacrifices and protecting its peices from potential captures.
But at the end game, its was terrible. It made very bad moves and lost.
At the end game, its took less than a second to decide on a move where it
took about 6 sec at the mid game. If we made the solver take its
time more, it may be able to make better desicions at the end game instead
of making quick moves that result in a loss. This makes sense since the
end game usually has very few captures and promotions and only a deeper
search will be able to see them.

Next try using a iterative deepening or breadth first search minimax variation
and limit it by time not depth. This should allow the solver to take its time
at the end game and make it search deeper into the game so it can see those
future captures and promotions.