Added iterative deepening to the solver so that it can continue searching until a time limit has been reached instead of only a depth limit. A small problem is still there. The 40 move rules in chechers says that if a game continues for 40 moves without a capture or promotion, then the game ends. Therefore, no matter how much time we want a move to use, the 40 move rule will limit us to only 40 plys without captures or promotions. This is still okay. The solver made decent moves, smart captures and defended its pieces.
One mistake I notices the solver make is, it move pieces as if
to make a sacrifice, but risks loosing more from the sacrifice.
It is not clear if these sacrifices are actually smart or not
but it seems like not. We should test the solver against a
human player to know for sure. Or maybe adjust the weight
involving holes. Make that weight negative so that the solver
knows holes are bad to have.

Also found and fixed a mistake in the alphabeta function. The
function was calling the minimax function instead of alphabeta.
This means the solver wasn't using alpha beta prunning before
now.