Found a very small but very bad bug in our Solver class. The bug was the reason why the solvers were making very obvious dumb decisions and not making very obvious smart decisions. The bug was causing the solver to make decisions based on searches that were incomplete for their ply. Since were using iterative depening for our decision making, the last iteration will very likely be incomplete because the time will run out before it finishes. ex: the solver finishes a 5 ply search and begins a 6 ply search. When the time runs out it returns a bluff move with some heuristic and if that heuristic is better than the 5 plys best move, then the solver will make the bluff move instead.
I fixed the bug after some long hours of debugging and tested the Solver1
solvers against one another, and they both made amazing moves. Especially
the solver using WEIGHTS_COMBO_5. It made very decisive moves, protected its
peices, promoted peices when it could detect them, made chain captures
whenever it could and also seemed to prevent letting its opponent make
captures and promotions in future moves.

One problem still persists for the solvers. When the solver detects
a promotion or capture after a passive move, it seems to procrastinate
and make the passive move instead. This is because it knows that in the near
future it will be able to make a promotion or capture even after making
the passive move, so it simply takes its time and makes passive moves
turn after turn.

A possible solution is to order the moves after every iteration. This way
the promotions and captures that are in the near future will be analyzed
and counted as best moves before the passive ones. I really think this
will solve the problem. Also it will make alpha beta prunning more
effective.